requirements.txt
tensorflow==2.15.0
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
umap-learn==0.5.5
matplotlib==3.8.3